v1.2 - more depth work, moriku in-engine, grass tilemaps?

+++ Soshikami - Mystery of Seasons/Assets/Scripts/PositionRenderSorter.cs
@@ -8,7 +8,7 @@ public class PositionRenderSorter : MonoBehaviour
     private int sortingOrderBase = 5000;
     private Renderer myRenderer;
     [SerializeField]
-    private int offset = 0;
+    private float offset = 0;
     [SerializeField]
     private bool runOnlyOnce = false;
